+++ webots_ros2_tests/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package webots_ros2_tests
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2023.1.0 (2023-XX-XX)
+------------------
+* Update webots_ros2_driver test to new WebotsController node.
+
 2023.0.4 (2023-05-23)
 ------------------
 * Added support for painted point clouds